+++ src/v1/MediaType/MediaType.spec.ts
@@ -35,7 +35,11 @@ import { expect } from "chai";
 import { describe } from "mocha";
 
 import { MediaType } from ".";
-import { InvalidMediaTypeExamples, ValidMediaTypeExamples } from "./MediaTypeExamples.spec";
+import {
+    InvalidMediaTypeExamples,
+    ValidContentTypeExamples,
+    ValidMediaTypeExamples,
+} from "./MediaTypeExamples.spec";
 
 describe("MediaType", () => {
     describe("constructor()", () => {
@@ -55,6 +59,27 @@ describe("MediaType", () => {
         }
     });
 
+    describe(".getContentType()", () => {
+        // tslint:disable-next-line: forin
+        for (const inputValue in ValidContentTypeExamples) {
+            it("correctly parses '" + inputValue + '"', () => {
+                const expectedValue = ValidContentTypeExamples[inputValue];
+                const unit = new MediaType(inputValue);
+
+                const actualValue = unit.getContentType();
+                expect(actualValue).to.eql(expectedValue);
+            });
+        }
+
+        it("returns the same value when called multiple times", () => {
+            const unit = new MediaType("application/vnd.tie-record+json");
+
+            const actualValue1 = unit.getContentType();
+            const actualValue2 = unit.getContentType();
+            expect(actualValue1).to.equal(actualValue2);
+        });
+    });
+
     describe(".parse()", () => {
         // tslint:disable-next-line: forin
         for (const inputValue in ValidMediaTypeExamples) {
@@ -66,6 +91,14 @@ describe("MediaType", () => {
                 expect(actualValue).to.eql(expectedValue);
             });
         }
+
+        it("returns the same value when called multiple times", () => {
+            const unit = new MediaType("application/vnd.tie-record+json");
+
+            const actualValue1 = unit.parse();
+            const actualValue2 = unit.parse();
+            expect(actualValue1).to.equal(actualValue2);
+        });
     });
 
     it("can be used as a string", () => {