+++ src/v1/MediaType/MediaType.ts
@@ -36,12 +36,25 @@ import { RefinedString } from "@ganbarodigital/ts-lib-value-objects/lib/v2";
 
 import { MediaTypeParts } from "./MediaTypeParts";
 import { mustBeMediaType } from "./mustBeMediaType";
+import { parseContentType } from "./parseContentType";
 import { parseMediaType } from "./parseMediaType";
 
 /**
  * value type. Represents RFC-compliant media type strings.
  */
 export class MediaType extends RefinedString {
+    /**
+     * internal cache. Stops us having to extract the content type
+     * from our value more than once
+     */
+    #contentType: string|undefined = undefined;
+
+    /**
+     * internal cache. Stops us having to parse our value
+     * more than once.
+     */
+    #parsed: MediaTypeParts|undefined = undefined;
+
     /**
      * smart constructor.
      *
@@ -53,9 +66,30 @@ export class MediaType extends RefinedString {
     }
 
     /**
+     * Gets the 'text/html' bit from 'text/html; charset=UTF-8' (for example)
+     */
+    public getContentType(): string {
+        // haven't we already done this?
+        if (!this.#contentType) {
+            // no, first time for everything!
+            this.#contentType = parseContentType(this.value, THROW_THE_ERROR);
+        }
+
+        // return our cached value
+        return this.#contentType;
+    }
+
+    /**
      * returns a breakdown of the individual components for this media type
      */
-    public parse(onError: OnError = THROW_THE_ERROR): MediaTypeParts {
-        return parseMediaType(this.value, onError);
+    public parse(): MediaTypeParts {
+        // haven't we already done this?
+        if (!this.#parsed) {
+            // no, first time for everything!
+            this.#parsed = parseMediaType(this.value, THROW_THE_ERROR);
+        }
+
+        // return our cached value
+        return this.#parsed;
     }
 }