+++ src/v1/MediaType/parseContentType.spec.ts
@@ -35,54 +35,34 @@ import { expect } from "chai";
 import { describe } from "mocha";
 
 import { MediaTypeMatchRegexIsBrokenError } from "../Errors/MediaTypeMatchRegexIsBroken";
-import { InvalidMediaTypeExamples, ValidMediaTypeExamples } from "./MediaTypeExamples.spec";
-import { parseMediaType, parseMediaTypeUnbound } from "./parseMediaType";
-import { MediaTypeParamRegex } from "./regexes";
+import { InvalidMediaTypeExamples, ValidContentTypeExamples } from "./MediaTypeExamples.spec";
+import { parseContentType, parseContentTypeUnbound } from "./parseContentType";
 
 // a valid regex, that doesn't populate named groups
 const BrokenMatchRegex = /^.*$/;
 
-describe("parseMediaType()", () => {
+describe("parseContentType()", () => {
     // tslint:disable-next-line: forin
-    for (const inputValue in ValidMediaTypeExamples) {
+    for (const inputValue in ValidContentTypeExamples) {
         it("correctly parses '" + inputValue + '"', () => {
-            const expectedValue = ValidMediaTypeExamples[inputValue];
-            const actualValue = parseMediaType(inputValue);
+            const expectedValue = ValidContentTypeExamples[inputValue];
+            const actualValue = parseContentType(inputValue);
 
-            expect(actualValue).to.eql(expectedValue);
+            expect(actualValue).to.equal(expectedValue);
         });
     }
 
     for (const inputValue of InvalidMediaTypeExamples) {
         it("rejects example '" + inputValue + "'", () => {
-            expect(() => parseMediaType(inputValue)).to.throw();
+            expect(() => parseContentType(inputValue)).to.throw();
         });
     }
 
     it("throws MediaTypeMatchRegexIsBrokenError if we use a garbage parser", () => {
         const inputValue = "text/plain";
-        expect(() => parseMediaTypeUnbound(
+        expect(() => parseContentTypeUnbound(
             BrokenMatchRegex,
-            MediaTypeParamRegex,
             inputValue,
         )).to.throw(MediaTypeMatchRegexIsBrokenError);
     });
-
-    // one day, I hope auto-conversion makes this possible!
-    //
-    // it("accepts a MediaType input type", () => {
-    //     const inputValue = "text/html; charset=UTF-8";
-    //     const expectedValue: MediaTypeParts = {
-    //         type: "text",
-    //         subtype: "html",
-    //         parameters: {
-    //             charset: "UTF-8",
-    //         },
-    //     };
-
-    //     const unit = new MediaType(inputValue);
-    //     const actualValue = parseMediaType(unit);
-
-    //     expect(actualValue).to.eql(expectedValue);
-    // });
 });